requirements.txt
rich
gitpython
requests
cython
sourcedefender